% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft_names.R
\docType{data}
\name{abbr2pft}
\alias{abbr2pft}
\alias{pft2abbr}
\title{PFT abbreviations}
\format{An object of class \code{character} of length 15.}
\usage{
abbr2pft

pft2abbr
}
\description{
Dictionary from PFT abbreviations to PFTs and vice-versa.
}
\keyword{datasets}
