% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm45_parameters.R
\docType{data}
\name{clm45_table81}
\alias{clm45_table81}
\title{CLM Table 8.1}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 5 columns.}
\usage{
clm45_table81
}
\description{
Community Land Model (CLM) version 4.5 parameters table (8.1), for 
comparison with results.
}
\details{
Columns are as follows:
 - `pft` -- Plant functional type, corresponding to analysis PFT
- `CN_leaf` -- Leaf C:N ratio, (g C g-1 N)
- `F_LNR` -- Fraction of leaf nitrogen in Rubisco (g N Rubisco g-1 N)
- `SLA` -- Specific leaf area (m2 g-1 C)
- `Vcmax_25` -- Vcmax at 25C at canopy top (mu~mol m-2 s-1)
}
\keyword{datasets}
