% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft_colors.R
\name{colortest}
\alias{colortest}
\title{Quick function for examining color schemes}
\usage{
colortest(cols)
}
\arguments{
\item{cols}{Character vector of colors}
}
\description{
Draws a bar plot [graphics::barplot()], with each bar shaded to the specified color.
Useful for quickly testing color schemes.
}
