% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_mv.R
\name{compare_mv}
\alias{compare_mv}
\title{Compare means and variances between multivariate and hierarchical models}
\usage{
compare_mv(summary_data, area_mass, pft_scheme)
}
\arguments{
\item{summary_data}{Nested tibble containing summary of results}

\item{area_mass}{Normalization, either "mass" or "area" (character)}

\item{pft_scheme}{PFT definition scheme (e.g. "clm45") (character)}
}
\value{
tibble containing plots comparing mean and variance
}
\description{
Compare means and variances between multivariate and hierarchical models
}
