% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_names.R
\name{drop_mass_area}
\alias{drop_mass_area}
\title{Drop the mass/area designation from trait, and return a clean factor}
\usage{
drop_mass_area(trait)
}
\arguments{
\item{trait}{Character or factor of traits}
}
\value{
Trait vector as factor, with levels [param_labels]
}
\description{
Drop the mass/area designation from trait, and return a clean factor
}
\author{
Alexey Shiklomanov
}
