% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-missing.R
\name{get_all_pairwise_missing}
\alias{get_all_pairwise_missing}
\title{Get all pairwise present and missing counts}
\usage{
get_all_pairwise_missing(trait_data)
}
\arguments{
\item{trait_data}{`data.frame` of data}
}
\value{
`data.frame` of number of pairwise and missing observations
}
\description{
Get all pairwise present and missing counts
}
\author{
Alexey Shiklomanov
}
