% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise-missing.R
\name{get_pairwise_missing}
\alias{get_pairwise_missing}
\title{Number of pairwise missing observations}
\usage{
get_pairwise_missing(xvar, yvar, trait_data)
}
\arguments{
\item{xvar, yvar}{(character) Columns to check for missingness}

\item{trait_data}{`data.frame` of data}
}
\value{
`data.frame` of missing and present observations
}
\description{
Number of pairwise missing observations
}
\author{
Alexey Shiklomanov
}
