% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft_colors.R
\docType{data}
\name{pft_colors}
\alias{pft_colors}
\title{PFT color scheme}
\format{An object of class \code{character} of length 14.}
\usage{
pft_colors
}
\description{
PFT color scheme
}
\keyword{datasets}
