% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft_colors.R
\docType{data}
\name{pft_inds}
\alias{pft_inds}
\title{Indices defining PFT colors}
\format{An object of class \code{numeric} of length 14.}
\usage{
pft_inds
}
\description{
Indices correspond to `tol21rainbow` color scheme.
}
\keyword{datasets}
