% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pft_colors.R
\docType{data}
\name{tol14rainbow}
\alias{tol14rainbow}
\alias{tol15rainbow}
\alias{tol18rainbow}
\alias{tol21rainbow}
\title{Paul Tol color schemes}
\format{An object of class \code{character} of length 14.}
\usage{
tol14rainbow

tol15rainbow

tol18rainbow

tol21rainbow
}
\description{
Effective color schemes for categorical data.
Available for 14, 15, 18, and 21 levels.
}
\keyword{datasets}
