% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_names.R
\docType{data}
\name{trait_labels}
\alias{trait_labels}
\alias{param_labels}
\alias{param_simple}
\alias{param_fancy_chr}
\alias{param_fancy}
\alias{param_markdown_nounit}
\alias{param_markdown}
\title{Trait labels}
\format{An object of class \code{character} of length 7.}
\usage{
param_labels

param_simple

param_fancy_chr

param_fancy

param_markdown_nounit

param_markdown
}
\description{
Various vectors for providing formatted trait labels for figures and tables.
}
\details{
Contains the following:
- `param_labels` -- Base trait names, without mass/area suffix
- `param_simple` -- R expression-formatted trait names with units, as character (used for `parse = TRUE`)
- `param_fancy_chr` -- R expression-formatted trait names with units, as character (used for `parse = TRUE`)
- `param_fancy` -- R expression-formatted trait names with units, as R `formula` vector
- `param_markdown_nounit` -- LaTeX-formatted trait names without units, as character
- `param_markdown` -- LaTeX-formatted trait names with units, as character
}
\keyword{datasets}
