% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait_names.R
\docType{data}
\name{trait_names}
\alias{trait_names}
\alias{mass_params}
\alias{area_params}
\alias{both_params}
\title{Trait names}
\format{An object of class \code{character} of length 7.}
\usage{
mass_params

area_params

both_params
}
\description{
Names of traits used in analyses, useful for factor levels and for quick 
selection from data frames.
}
\details{
Contains the following:
- `mass_params` -- Mass-based traits
- `area_params` -- Area-based traits
- `both_params` -- "Base" names
}
\keyword{datasets}
