% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try-data.R
\name{try_data}
\alias{try_data}
\title{Load TRY data from `extdata`}
\usage{
try_data(mass_area = NULL, f = here::here("extdata",
  "data_anonymized.csv"))
}
\arguments{
\item{mass_area}{(Optional) Retrieve only `"mass"` or `"area"` normalized
traits}

\item{f}{Path to data file. Default is `here(extdata/data_anonymized.csv)`}
}
\value{
`tibble` containing a `pft` column and the trait data (one trait per
  column)
}
\description{
Load TRY data from `extdata`
}
\author{
Alexey Shiklomanov
}
